feature/payment

+++ app/src/main/java/com/example/e_social/models/data/response/DataX.kt
@@ -14,5 +14,5 @@ data class DataX(
     val user_id: String,
     val username: String,
     val type: String,
-    val accountBalance: Int
+    val accountBalance: Long
 )
